day layout algorithm css customizable
& move Event class into overlap algorithm
